Updated openstack/openstack Project: openstack/nova  a0fcff8badc6d369307c1f2cbac958514b0ecbc5
Py3: fix serial console output

The compute API expects the serial console output to be a string,
attempting to use a regex to remove some characters. This will fail
as we are returning a byte array.

Also, since the API is expected to return the console output as a
string, the compute nodes may just return a string as well.

Closes-Bug: #1591996
Partially implements blueprint: nova-python3-newton

Change-Id: I5d3097f1d30f3b3568a5421e0d68aaf0797c850a