Updated openstack/openstack Project: openstack/nova  7c9ad8182cd39a2ceb82e1619a7bfd5203dfde7d
Don't raise error when filtering on custom metadata

Hosts can have custom metadata. There is no restriction on the key names
used in this metadata, so we should not be raising an exception when
checking for the existence of any metadata key.

Originally worked on: https://review.openstack.org/#/c/271401

Co-Authored-By: Xiaowei Qian <xiaowei.qian@easystack.cn>

Closes-Bug: #1537062

Change-Id: Ie5ff3c1847e9c4533822a77d443e4ce1fcf047fe