Updated openstack/openstack Project: openstack/tempest  1fafc756a069ceb0988179e9c16e8d863c1b5582
Remove DataGenerator from admin V2 identity tests

The identity tests have a DataGenerator that handles resource creation
and cleanup for tests, however using this DataGenerator has proven to
make debugging more difficult. The object storage tests used to use
this DataGenerator as well but was removed for the same reason. This
is the first of a series of commits to removes the DataGenerator from
the Identity tests. This commit removes the DataGenerator from the
admin v2 tests and moves resource creation and cleanup back to the
tests or the resource_setup, resource_cleanup methods.

Change-Id: I2767d962826d0a6249b14afb325cbaad7fcaa34c