Updated openstack/openstack Project: openstack/openstack-manuals  c2ae2491cd83dd114478634dc9830d774d3556f8
Correct note on how Horizon modifies flavors

As noted in the bug below, Horizon doesn't actually
modify flavors, it deletes them and creates a new one
with the same name and modified values.

This patch makes this clear in the documentation so
that the behavioral side effects are clear for users
(such as id-regeneration).

Change-Id: Ieebdf1809149afbd1b4fc27ceef31d6483bfcd88
Closes-Bug: #1519573