Updated openstack/openstack Project: openstack/nova  e6deb676cdeb630add92663cc59f0680e38ce2a5
Use flavor attributes instead of deprecated instance resources

We have almost completely moved to using the nested flavor on
instance objects when we need to determine things like root disk
size and memory usage. However the legacy instance.$resource fields
still exist and are used by the core resource tracking and claims
code. This code is hairy and was never updated after the flavor
work because, well, it's not fun. With the coming work to change
the resource management with pools, it's time to convert these
last few vestigal uses so we can deprecate those older fields and
remove them at some point.

This removes almost all of the remaining usage, with the exception
of the simple-tenant-usage extension, which will be converted
separately for brevity.

Change-Id: I316b16b224d490f03c55a0f85f98d70ab05a4375