Updated openstack/openstack Project: openstack/heat  2303d2e2c1478011aa2df6e9b9527aa1dabc8cba
Implement the correct total ordering for scheduler.Timeout

The commit 640abe0c12e63c207fcf67592838f112a29f5b43 implemented rich
comparison methods for the scheduler.Timeout class in order to be
compatible with Python3. However, it left out most of the checks to ensure
that Timeouts are compared only to other timeouts. It also left behind a
completely incorrect implmentation of __cmp__, although it didn't matter
because that is ignored for objects that implement rich comparisons.

This change deletes __cmp__, adds the type checking, and uses
functools.total_ordering to cut down on the amount of boilerplate.

Change-Id: Idf65d41705cefe9dab8ad5a518b142314077086b