Updated openstack/openstack Project: openstack/ceilometer  97a24d4a9a544d990e7f5cdc9ff7ffea6f720cad
gnocchi: use events to end Gnocchi resource

This change introduces a way to handle resources with the events
subsystem.

Also this need to be supported for type of resource.
To support it, etc/ceilometer/gnocchi_resources.yaml need to be
updated to add which event creates/updates/deletes the resource and
the mapping between event traits and gnocchi resource attributes.

This change adds the code to handle and the support for the 'instance'
and 'image' resource.

Only the delete event is support for now.

Related-bug: #1483634
Change-Id: Icd77137a74bccb5b2be078f206f153f0e9aa86c5