Updated openstack/openstack Project: openstack/ironic-python-agent  ad60806f93c7d4cc9eec61ff24486db091ecb215
Small refactor in the root device loop matching logic

When matching the root device hints passed to Ironic and the devices
present in the disk, the logic to the "size" attribute was different was
outside the main loop. This patch refactors that uses the same loop for
matching the size attribute, improving a little the readability.

This patch also makes sure that we convert the value of "size" to
integer before attempting to match it.

Change-Id: I619153232b9b58696369185ad5be90ccfd38a4ed