Updated openstack/openstack Project: openstack/tempest  1c984e4034a0058dbb99b780799de735a06e46e3
Remove DataGenerator from the dentity base

The identity tests have a DataGenerator that handles resource creation
and cleanup for tests, however using this DataGenerator has proven to
make debugging more difficult. The object storage tests used to use
this DataGenerator as well but was removed for the same reason.

This is the third and last of a series of commits to removes the
DataGenerator from the Identity tests. This commit removes the
DataGenerator class from the identity base module.

Closes-Bug: 1280333
Change-Id: I4e2a99767e74da11f0cb420a3948900466bf50d3