Updated openstack/openstack Project: openstack/neutron-lbaas  26a5398e823d27f13a4707d201ae504777cada77
Enable removal of SSL/TLS reference on listener

To remove TLS, a person would update the listener with
default_tls_container_ref="" or None and the expected result is for
the default_tls_container_ref to be empty but instead it was filled
with the current value in the database.

This change fixes the if statements that was responsible for this bug.

Change-Id: I9c8610998743090297ec6cbc76d766be5c13d624
Closes-Bug: #1569536