Updated openstack/openstack Project: openstack-infra/zuul  d085d2d597aba3f7fadf535c846c9b0513991911
Use fetch instead of update in merger

According to a well written and therefore almost certainly correct
answer on SO[1], fetch and update are nearly the same, especially
when used in the context of a single remote (as we are here).

Since we sometimes want correct tag information as well as correct
branch information for our repos, switch from update to fetch, with
the '--tags' argument.

[1] http://stackoverflow.com/questions/2688251/what-is-the-difference-between-git-fetch-origin-and-git-remote-update-origin

Change-Id: I3a31b84e7d65efa5430fb29a77047ae4b16ad18c