Updated openstack/openstack Project: openstack/cinder  5ba925270d5f65a4ddc548d6ba59281b1c052e65
Storwize: Fix exception ProcessExecutionError

There is specific code in storwize_svc_common
get_host_from_connector to check for CMMVC5754E in
the exception coming back from lshost(), but the exception
type was ProcessExecutionError. I believe that was correct
at one point in the past, but it has been switched to be
exception.VolumeBackendAPIException now so this change looks
for that. Adds UT and other minor cleanup.

Change-Id: If3220e88ae191ca3b5ff0949212e2f23552de300
Closes-Bug: #1578294