Updated openstack/openstack Project: openstack/keystone  18a1f1a554bb38114b24116ffbb3a1c38900f816
Use freezegun to increment the clock in test_v3_filters

In preparation for making Fernet the default token provider, we must freeze and
increment the clock in certain tests. This is because Fernet does not support
sub-second precision and if a Fernet token is validated in the same second that
a revocation event has occured, it will be considered invalid.

This commit makes it so that we freeze the clock in
test_list_users_filtered_by_funny_name() and increment it by one second when we
go to list users by name.

This fix was originally a part of https://review.openstack.org/#/c/258650 but
this is an attempt to break 258650 into smaller, more reviewable, pieces.

Partial-Bug: 1561054
Change-Id: I3ea4af702e6914775156c605799992a41e8e75c5