Updated openstack/openstack Project: openstack-infra/nodepool  03debb7b82308f1d1808fd83ecf2189717056822
Cache function worker map

We need the function worker map to keep track of how queued functions
relate to label demand in nodepool but the data required for this can
grow quite large so rather than check it on every loop through the
nodepool run loop cache the info unless we find an unknown function or
the ttl expires.

By default the ttl is one day but can be configured in the nodepool yaml
configuration.

Change-Id: Id069bdc5c35d6c9ce8f57af5ae2d8ea840b1e0d9