Updated openstack/openstack Project: openstack/keystone  85b00a60f3fd8f22d2be9d5f047ec4a13cce7620
Prevent error when duplicate mapping is created

The mapping has unique constraint and creating a mapping with the same
local entity will fail. Creation of mapping with the same local entity
might happen because of race condition, that would happen if 2 processes
start mapping of the same local entity at the same time.

Return the existing public_id if creation failed due to duplicate.

Change-Id: Ic6e6ec9614e16a49c791dc6c06a4464d18be09be
Related-Bug: 1582585