Updated openstack/openstack Project: openstack/nova  25187117a383c39a591bf18be6d6ba092238ecf9
Improve the help text for the network options (4)

The previous commit moved all the options from nova/network/manager.py
to a central location in nova/conf/network.py. This commit adds
additional help text to those options. Because of the large number of
options, the improved help text will be split into several patches; this
is patch 4 of 4.

Blueprint centralize-config-options-newton

Change-Id: I06d9a76afe6a926db5f3f689b212f2562dbb6a92