Updated openstack/openstack Project: openstack/trove  342ffd78c2046673fc8dad4bf16c4b63fe5d9b39
Serialize py34 tests

It turns out that the way the python34 tests are run in parallel is
to spin up several processes (not threads).  This means that putting
locks around creating the database in a single 'run' would need an
inter-process lock not a thread lock.

While the feasibility of this is investigated, change the behaviour
of the tests to run in series to increase stability.

On a 4 cpu test machine this only increased the run time by 70s.

Partially implements: blueprint trove-python3

Change-Id: I530fcf22b697148a3d0a286015c5a79481528d8e