Updated openstack/openstack Project: openstack/trove  79006ce1630f2cc4290e8081bba7faf3748f98d7
Support newer mysql in tests

Newer mysql cli clients emit a warning when you put a password on the
cli. This warning was being interpreted as a failure by the test
framework. Change the framework to accept the warning or empty string as
valid responses when running the mysql cli client to support using the
cli client in the trove tests.

I am not entirely sure when this new behavior arrived in mysql but it is
present in Ubuntu Xenial's Mysql 5.7 packages.

Change-Id: Ia148d72a7646f4c424139232d6388ff14d1d8686