Updated openstack/openstack Project: openstack/nova  02846f26f55464e9a728bb8c451afe2c9eed9626
Make libvirt driver use flavor fields instead of legacy ones

This makes the libvirt driver use instance.flavor.root_gb instead
of the legacy instance.root_gb field (and friends).

Also, this fixes a bunch of bugs in tests where the incorrect
value of memory in megabytes was being compared to the actual
result in kilobytes from the config generation code. This was
not detected before because of multiply-by-zero.

Change-Id: I23415edb0347741690fe8fcc75cdc2f3d83be4fc