Updated openstack/openstack Project: openstack/ironic  e0a979982e916fc9e809e098c85ebb568ca1ba00
Add and document the "rotational" root device hint

This patch is adding a new root device hint called "rotational". This
hint is used to identify whether a device is rotational or not making it
easy to distinguish HDDs and SSDs when choosing which disk Ironic
should deploy the image onto.

Closes-Bug: #1599517
Depends-On: I270fe57df825929bdef7911b3a6757cf7163a5f1
Change-Id: Id630a0b9d02ed8e1bd674c32bef0d489849c3f29