Updated openstack/openstack Project: openstack-infra/irc-meetings  5c9e5e670e8a1b9d28e19afca56639a1bea771ab
Set meeting chairs to the canonical form

We've asked for quite some time that meeting chairs be of the form
Full Name (IRC_Nick) with multiple chairs separated by a ','.

This patch reduces the debt here by filling in those details from:
 www.openstack.org/community/members/

It doesn't complete the task as some meeting chairs do NOT have matching
entries in that database.

Also no effort has been made to verify that the listed meeting chair is
in fact the regular meeting chair.

Change-Id: If1abfb43c8ae6f1cd8d3f676c8e5131e3937e86a