Updated openstack/openstack Project: openstack/keystone  8dfb8747241d38332899e57a3010fe886768409c
Run AuthTokenTests against fernet and uuid

Previously, AuthTokenTests inherited the test structure and ran the tests with
the default keystone configuration. This commit breaks AuthTokenTests to
inherit from `object` and sets up FernetAuthTokenTests and UUIDAuthTokenTests
to setup configuration appropriately and run the tests.

This fix was originally a part of https://review.openstack.org/#/c/258650 but
this is an attempt to break 258650 into smaller, more reviewable, pieces.

Co-Authored-By: Raildo Mascena <raildo@lsd.ufcg.edu.br>
Co-Authored-By: Adam Young <ayound@redhat.com>

Change-Id: I9acacfe7db4997add5505a9ee1972139af11979e
Partial-Bug: 1561054