Updated openstack/openstack Project: openstack/heat  ce4c7cc7cece2fb1eb94bf11bea096ced7cfc8e6
Unit tests: add a test resource that takes multiple steps

All of the current test resource types complete in a single step (i.e. have
no check_*_complete() methods). In order to test longer-running resources,
add a resource type in which the number of steps can be specified as a
property.

Change-Id: Ia894efa3207130a187032f2f667c096dee883aea