Updated openstack/openstack Project: openstack/heat  94a8d440c6de1d59493c2bcc0de3fd3022d98367
Scheduler: Add a progress callback to TaskRunner

This allows the caller to specify an arbitrary function to be called after
each sleep() when running a task. Callers can use this to keep track of
progress or keep track of external state. Any exception raised by the
callback function will be thrown into the task.

Change-Id: I637166f9448169eed8696e5b0fba106b8de1547a