Updated openstack/openstack Project: openstack/heat  b01f63271bc124fe1399889d268d015449bac9c3
Use a queue instead of an event to send messages

This has a couple of effects. One is that we can send more than one message
to an action (since the action need not terminate immediately upon
receiving a message). The other is that we only ever receive the message
once, instead of every time we check for it once it has been sent.

Change-Id: Ib3b33534c0e8b03dfbbb1acde6f7b2d3dcfa82b6