Updated openstack/openstack Project: openstack/heat  d78569b29e5a28427ca53bce3b14c71b8ea833f9
Allow a grace period when cancelling a stack update

Use a progress callback for cancelling updates so that we will apply the
usual grace period to allow resources to complete when cancelling a stack
update, instead of cancelling immediately. Previously, resources could
unneccessarily be marked as FAILED, resulting in them unneccessarily being
replaced on a subsequent update.

Change-Id: I0a5bf998be764e4d6d4c6ea86aaca7f715b93d06
Related-Bug: #1591337