Updated openstack/openstack Project: openstack/tooz  c3460f6d3657165b8e18bd06b724e484ae869e37
Share _get_random_uuid() among all tests

Currently, _get_random_uuid() is defined in test_coordination.py as a
staticmethod. This in turn forces multiple test suites to either use
that function from test_coordination (which was not done), or
re-implement the function on each call (which was done).

This patch proposes a move of the said function to tests/__init__.py,
which will allow every module to simply invoke
'tooz.tests.get_random_uuid()' if they wanted a random uuid. This will
eliminate a lot of code duplication in the test suite.

Change-Id: Idfc50d9da23790613d30bd87fb4da8eb272dce1f