Updated openstack/openstack Project: openstack/horizon  97123bdd93274a660ae907b5a2da98aef3a8beaf
Image uses hz-property for its drawer information

The Image drawer should use the hz-property so it is kept in sync with all other
uses of the properties (in tables, etc.).

Gets rid of filters in favor of simple mappings or functions.

Moves several functions out of the module into a basic images-service so
it's easy to access/test functions.

Change-Id: I4d881744a398f9aa076830b576cb9f66335d8ba9
Partially-Implements: blueprint angular-registry