Updated openstack/openstack Project: openstack/nova  4e4adb5039703ce0c372af2d15b60a52428557e5
Stop instance build if BuildRequest deleted

As part of the cellsv2 instance boot process there is a check that the
BuildRequest still exists before sending the request to a compute. This
is how an instance delete for an unscheduled booting instance will stop
the boot from proceeding.

However if there are some older nova-api services then the BuildRequest
may not have ever existed so it's non-existence does not indicate a
delete occurred. A service version check accounts for this and lets the
build proceed as normal. In that case the delete code will also need to
account for the lower service version and clean up the instance from the
cell.

Change-Id: Id65ab1911071608a739bea96e071939baa388493
Partially-implements: bp add-buildrequest-obj
Partially-implements: bp cells-scheduling-interaction