Updated openstack/openstack Project: openstack/swift  45080a5177354e6329a62624d384fcddb8ba0031
Prevent CPU spinning when there are no children

If you deploy an object server but have no rings at all (and are using
servers-per-port), then the CPU will spin as it checks for child
processes since there are actually no child processes to check.

This patch adds a sleep so that the CPU doesn't spin.

Change-Id: Iece62367aa2481a21752144b1f4477a3713282fe