Updated openstack/openstack Project: openstack/nova  849d5a2f80d2ca481321acb900237c4574143244
Update binding:profile for SR-IOV ports

The libvirt driver relies on the network information for configuring the
SR-IOV interfaces (binding:vnic_type). The network information is specified
when creating a Neutron port (binding:profile - which contain the PCI address
of the device).

During a migration, the Neutron port information needs to be updated based on
the allocated PCI device on the migrated node.

Co-Authored-By: Moshe Levi <moshele@mellanox.com>

Closes-Bug: #1512880

Change-Id: I7423907ef33648669decc561fc3461415b877ba6