Updated openstack/openstack Project: openstack/nova  414df1e56ea9df700756a1732125e06c5d97d792
libvirt: update guest time after suspend

When an instance is resumed from suspension or a migration, the guest's
clock might get a large delay between the current and the suspension time.
In some cases, NTP might not be able to resynchronize the guest time and
will result in a skewed clock.

In order to address the issue, this patch will try to set the guest time
after it is resumed. This operation will require the guest agent to be
configured and running in order to be able to run. However, this operation
will not be disruptive.

This code is similar to a change that has been commited to oVirt.
https://gerrit.ovirt.org/#/c/50426

Closes-Bug: #1593745
Change-Id: I2732af50aab238502ccdc067f62290fd962c8c64