Updated openstack/openstack Project: openstack/requirements  8f1ccaf70f3c4aa6aa53fbddf589310c02eb4ce0
Pin to a lower version of store release

Due to some translation changes not being formatted correctly, the
latest glance_store release is resulting into glance docs gate to fail.

This is blocking reviews on glance and we don't have a concerete
solution on the underlying changes in oslo.config/oslosphinx or
elsewhere required.

This is the best attempt to unblock glance reviews.

Change-Id: I9f51741e1ce1d37b9d6fd1c58c6754317d25ced9
ref: http://lists.openstack.org/pipermail/openstack-docs/2016-July/thread.html#8899