Updated openstack/openstack Project: openstack-infra/release-tools  9a80b51b8c631ad9d8640cbd763af33872a026ae
force a fuller update of the repo being cloned for tagging

zuul-cloner may be using an out of date cache, and may have pointed
originally at a repo on a zuul merger that doesn't have all tags. Force
an update of all remotes and fetch all tags to make sure they are
present before we start trying to add any new ones.

Change-Id: I29cdae3ff8195e4b7e35ac486429394f17919bff
Signed-off-by: Doug Hellmann <doug@doughellmann.com>