Updated openstack/openstack Project: openstack/heat-translator  7d83fcc31d2e1196efe18c99a7c701d10f367187
use keystoneauth and OpenStack clients

The parser should not rely on the OS_* environment variables for
authentication, as it will fail with more complex setups, when using the
"clouds.yaml" configuration file or when being called through osc with
arguments instead of environment variables.

Moreover, the client should not use "requests" directly to do GETs to
access the nova flavors and glance images, but it should use the
existing clients.

When used through the python-openstackclient it is possible to force the
client to request authentication. The authentication will be handled by
the client, so that the parser does not need to worry about it (so that
authentication plugins can be used). Then, the existing authentication
session can be reused for all the client interaction with OpenStack to
fetch the flavor and images information.

Change-Id: I9f65e7d46686c37bb44ef18756ebd295ee4961de