Updated openstack/openstack Project: openstack/neutron  55c4b9afc2d499e4163e9380f7ef6ba3cf61d0bf
Add function to return all hosts with mapped segments

In the agentless topology, for example ovn, hosts might be removed from
the topology, during neutron service down. When neutron service up, a
full sync will be performed. But only existing hosts will be synced.
There is no way to know which hosts are removed, and there are stale
SegmentHostMapping left behind.

This patch adds a function to return all hosts with mapped segments, so
that it is possible to compare existing hosts with all hosts in DB. And
delete stale records if there is any. It is applied at [1].

[1] https://review.openstack.org/337685

Change-Id: I53863c9983e933872fb0adbc964e5a88d0238bf2
Partially-implements: blueprint routed-networks