Updated openstack/openstack Project: openstack/neutron  9fd6b5841d4d9e506812587f7d8b82868d267674
Fixes port device_id/device_owner change in failed operation

If a router is a attached to a neutron port and that operation
fails at a late stage in the processing then the 'device_id' and
'device_owner' attributes of the port are still changed. The
failed operation should not change them.

The reason is, as of commit 0947458018725b241603139f4ec6f92e84b2f29b,
the update of those attributes happens outside of the DB transaction.
For failed operation, the values are not reset back to their original
values.

Change-Id: I797df266dafc41843408dc95a6ce9f986db5c21c
Closes-Bug: #1523859