Updated openstack/openstack Project: openstack/horizon  8d58c6f3c4868a484ad3b380dc8c0f75ec8bb238
Cleanup the toastService in *.spec.js files

We should not use toastService to test if an API call
went through successfully. This is extraneous.

This patch was motivated by the fact that we want to
remove conf.js, so we are moving the toast config
options to toast.service.js.

Change-Id: I966fb0d492ae8e9ae70af2882655cd36b4214a4a