Updated openstack/openstack Project: openstack/tripleo-heat-templates  094c9d0bf6f697c32ad6c984161c53a6c4b38614
Fixes type and description for NeutronL3HA

puppet-neutron takes this variable as boolean.  Although it doesn't
change the behavior in master (because the variable is used directly as
config), in mitaka it is used as a conditional and should be properly
fixed.

Also a fix is needed in python-tripleoclient because it is hard coded to
be True there based on number of controllers being greater than 1.
Therefore we shoudl remove that logic from tripleoclient and implement
it in THT.  In order to do that the pacemaker version of the variable is
defaulted to true.

Partial-Bug: 1605379

Change-Id: I0b797dbe188382e2dc32506913aaa60a0a21bd68
Signed-off-by: Tim Rozet <trozet@redhat.com>